+++ packages/dart_suite/lib/src/extensions/record_extensions.dart
@@ -1,1 +1,0 @@
-extension RecordExtensions on (int, int) {}