+++ packages/dart_suite/lib/src/utility.dart
@@ -1,6 +1,6 @@
 export 'utility/validator_mixin.dart';
 export 'utility/guard.dart';
-export 'utility/typedef.dart';
+export 'typedefs/typedefs.dart';
 export 'utility/re_case.dart';
 export 'utility/range.dart';
 export 'utility/repository_base.dart'; // Add in v0.0.3
@@ -11,3 +11,5 @@ export 'async_control/debounce.dart';
 
 export 'timeago/timeago.dart';
 export 'regpattern/regpatterns.dart';
+
+export 'dsa/lru_cache.dart';