+++ packages/dart_suite/test/main_test.dart
@@ -4,6 +4,7 @@ import 'src/timeago_test.dart' as timeago_test;
 import 'src/retry_test.dart' as retry_test;
 import 'src/debounce_test.dart' as debounce_test;
 import 'src/throttle_test.dart' as throttle_test;
+import 'src/lru_cache_test.dart' as lru_cache_test;
 
 void main() {
   guard_test.main();
@@ -12,4 +13,5 @@ void main() {
   retry_test.main();
   debounce_test.main();
   throttle_test.main();
+  lru_cache_test.main();
 }